Initial

+++ commands/say.js
@@ -1,6 +1,7 @@
 const Discord = require('discord.js');
 
 exports.run = (bot, message, args) => {
+    if(message.channel.type === "dm") return message.channel.send("Sorry, but using this command in a dm will break my code.")
     let shotomessage = args.join(" ");
     if(!shotomessage) return message.channel.send("You can't tell me to say nothing. Try again.")
     message.delete().catch();
